requirements.txt
numpy
matplotlib

folium
geopy
jupyter
Pillow